# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trà Đá Mentor Hub API",
    "description": "REST API for mentoring program operations. Authenticated routes require `Authorization: Bearer <accessToken>`. List endpoints return a raw array unless `page` is set, in which case the body is `{ data, page, limit, total }`. Errors use `{ success: false, code, message }`.",
    "version": "1.1.0",
    "contact": {
      "name": "Trà Đá Mentor",
      "url": "https://github.com/TheHien04/Tra-Da-Mentor-Hub"
    }
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Local development" },
    { "url": "/", "description": "Current host (production single-origin deploy)" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "Q": { "name": "q", "in": "query", "schema": { "type": "string" } },
      "Page": { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
      "Track": { "name": "track", "in": "query", "schema": { "type": "string" } },
      "MentorshipType": {
        "name": "mentorshipType",
        "in": "query",
        "schema": { "type": "string", "enum": ["GROUP", "ONE_ON_ONE"] }
      },
      "Duration": {
        "name": "duration",
        "in": "query",
        "schema": { "type": "string", "enum": ["SHORT_TERM", "MEDIUM_TERM", "LONG_TERM"] }
      },
      "Capacity": {
        "name": "capacity",
        "in": "query",
        "schema": { "type": "string", "enum": ["active", "full", "none"] }
      },
      "Expertise": { "name": "expertise", "in": "query", "schema": { "type": "string" } },
      "ApplicationStatus": { "name": "applicationStatus", "in": "query", "schema": { "type": "string" } },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["success", "code", "message"],
        "properties": {
          "success": { "type": "boolean", "example": false },
          "code": {
            "type": "string",
            "enum": [
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "VALIDATION",
              "CONFLICT",
              "EMAIL_TAKEN",
              "SLOT_TAKEN",
              "SLOT_CONFLICT",
              "SLOT_PAST",
              "INVALID_CREDENTIALS",
              "ACCOUNT_INACTIVE",
              "EMAIL_UNVERIFIED",
              "DEMO_ACCOUNT",
              "FILE_TOO_LARGE",
              "INTERNAL"
            ]
          },
          "message": { "type": "string" }
        }
      },
...
      "Mentor": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "userId": { "type": "string", "nullable": true },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "track": { "type": "string" },
          "bio": { "type": "string" },
          "expertise": { "type": "array", "items": { "type": "string" } },
          "maxMentees": { "type": "number" },
          "mentees": { "type": "array", "items": { "type": "string" } },
          "groups": { "type": "array", "items": { "type": "string" } },
          "mentorshipType": { "type": "string", "enum": ["GROUP", "ONE_ON_ONE"] },
          "duration": { "type": "string", "enum": ["SHORT_TERM", "MEDIUM_TERM", "LONG_TERM"] },
          "avatarUrl": { "type": "string" }
        }
      },
      "MentorWrite": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "track": { "type": "string" },
          "bio": { "type": "string" },
          "expertise": { "type": "array", "items": { "type": "string" } },
          "maxMentees": { "type": "number" },
          "mentorshipType": { "type": "string" },
          "duration": { "type": "string" },
          "avatarUrl": { "type": "string" }
        }
      },
      "PagedMentors": {
        "type": "object",
        "required": ["data", "page", "limit", "total"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Mentor" } },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
      "Mentee": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "userId": { "type": "string", "nullable": true },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "track": { "type": "string" },
          "school": { "type": "string" },
          "interests": { "type": "array", "items": { "type": "string" } },
          "goals": { "type": "array", "items": { "type": "string" } },
          "progress": { "type": "number" },
          "mentorId": { "type": "string", "nullable": true },
          "groupId": { "type": "string", "nullable": true },
          "mentorshipType": { "type": "string", "enum": ["GROUP", "ONE_ON_ONE"] },
          "applicationStatus": { "type": "string" },
          "avatarUrl": { "type": "string" },
          "mentor": { "type": "string" },
          "group": { "type": "string" }
        }
      },
      "MenteeWrite": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "track": { "type": "string" },
          "school": { "type": "string" },
          "interests": { "type": "array", "items": { "type": "string" } },
          "goals": { "type": "array", "items": { "type": "string" } },
          "progress": { "type": "number" },
          "mentorId": { "type": "string" },
          "groupId": { "type": "string" },
          "mentorshipType": { "type": "string" },
          "applicationStatus": { "type": "string" },
          "avatarUrl": { "type": "string" }
        }
      },
...
      "PagedMentees": {
        "type": "object",
        "required": ["data", "page", "limit", "total"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Mentee" } },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
      "MeetingSchedule": {
        "type": "object",
        "properties": {
          "frequency": { "type": "string" },
          "dayOfWeek": { "type": "string" },
          "time": { "type": "string" }
        }
      },
      "Group": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "topic": { "type": "string" },
          "mentorId": { "type": "string" },
          "mentor": {
            "type": "object",
            "properties": {
              "_id": { "type": "string" },
              "name": { "type": "string" },
              "email": { "type": "string" }
            }
          },
          "mentees": { "type": "array", "items": { "type": "string" } },
          "maxSize": { "type": "number" },
          "meetingSchedule": { "$ref": "#/components/schemas/MeetingSchedule" },
          "frequency": { "type": "string" },
          "dayOfWeek": { "type": "string" },
          "time": { "type": "string" }
        }
      },
      "GroupWrite": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "topic": { "type": "string" },
          "mentorId": { "type": "string" },
          "mentees": { "type": "array", "items": { "type": "string" } },
          "maxSize": { "type": "number" },
          "meetingSchedule": { "$ref": "#/components/schemas/MeetingSchedule" }
        }
      },
      "PagedGroups": {
        "type": "object",
        "required": ["data", "page", "limit", "total"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Group" } },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
      "Activity": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "type": { "type": "string" },
          "message": { "type": "string" },
          "description": { "type": "string" },
          "createdAt": { "type": "string" },
          "timestamp": { "type": "string" }
        }
      },
      "Slot": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "mentorId": { "type": "string" },
          "menteeId": { "type": "string", "nullable": true },
          "date": { "type": "string" },
          "time": { "type": "string" },
          "duration": { "type": "number" },
          "meetingLink": { "type": "string" },
          "bookedBy": { "type": "string", "nullable": true },
          "mentorName": { "type": "string", "nullable": true },
          "menteeName": { "type": "string", "nullable": true },
          "googleCalendarEventId": { "type": "string", "nullable": true }
        }
      },
      "SlotWrite": {
        "type": "object",
        "required": ["mentorId", "date", "time", "duration"],
        "properties": {
          "mentorId": { "type": "string" },
          "date": { "type": "string" },
          "time": { "type": "string" },
          "duration": { "type": "number" },
          "meetingLink": { "type": "string" }
        }
      },
...
      "SessionLog": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "mentorId": { "type": "string" },
          "menteeId": { "type": "string" },
          "sessionDate": { "type": "string" },
          "topic": { "type": "string" },
          "mentorScore": { "type": "number", "nullable": true },
          "menteeScore": { "type": "number", "nullable": true },
          "mentorNeedsSupport": { "type": "boolean" },
          "mentorSupportReason": { "type": "string", "nullable": true },
          "menteeNeedsSupport": { "type": "boolean" },
          "menteeSupportReason": { "type": "string", "nullable": true },
          "completedByMentor": { "type": "boolean" },
          "completedByMentee": { "type": "boolean" }
        }
      },
      "SessionLogWrite": {
        "type": "object",
        "required": ["mentorId", "menteeId", "sessionDate", "topic"],
        "properties": {
          "mentorId": { "type": "string" },
          "menteeId": { "type": "string" },
          "sessionDate": { "type": "string" },
          "topic": { "type": "string" },
          "mentorScore": { "type": "number" },
          "menteeScore": { "type": "number" },
          "mentorNeedsSupport": { "type": "boolean" },
          "mentorSupportReason": { "type": "string" },
          "menteeNeedsSupport": { "type": "boolean" },
          "menteeSupportReason": { "type": "string" },
          "completedByMentor": { "type": "boolean" },
          "completedByMentee": { "type": "boolean" }
        }
      },
      "Notification": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "userId": { "type": "string" },
          "title": { "type": "string" },
          "message": { "type": "string" },
          "type": { "type": "string", "enum": ["info", "success", "warning", "error"] },
          "href": { "type": "string", "nullable": true },
          "read": { "type": "boolean" },
          "createdAt": { "type": "string" }
        }
      },
      "NotificationList": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Notification" } }
        }
      },
      "Invite": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "email": { "type": "string" },
          "role": { "type": "string" },
          "createdAt": { "type": "string" },
          "expiresAt": { "type": "string" },
          "usedAt": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["active", "expired", "used"] },
          "link": { "type": "string" },
          "success": { "type": "boolean" }
        }
      },
      "InviteWrite": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": { "type": "string" },
          "role": { "type": "string", "enum": ["mentor", "mentee", "admin"] }
        }
      },
      "InviteList": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Invite" } }
        }
      },
...
      "Testimonial": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "menteeName": { "type": "string" },
          "mentorName": { "type": "string" },
          "content": { "type": "string" },
          "rating": { "type": "number" },
          "track": { "type": "string", "enum": ["career", "personal", "soft_skills"] },
          "date": { "type": "string" },
          "status": { "type": "string", "enum": ["PUBLISHED", "PENDING", "REJECTED"] }
        }
      },
      "TestimonialWrite": {
        "type": "object",
        "required": ["menteeName", "mentorName", "content", "rating", "track"],
        "properties": {
          "menteeName": { "type": "string" },
          "mentorName": { "type": "string" },
          "content": { "type": "string" },
          "rating": { "type": "number" },
          "track": { "type": "string", "enum": ["career", "personal", "soft_skills"] },
          "status": { "type": "string", "enum": ["PUBLISHED", "PENDING", "REJECTED"] }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/invites": {
      "post": {
        "tags": ["Invites"],
        "summary": "Create invite (admin)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InviteWrite" } } }
        },
        "responses": {
          "201": {
            "description": "Invite created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invite" } } }
          }
        }
      },
      "get": {
        "tags": ["Invites"],
        "summary": "List invites (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Invite list",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InviteList" } } }
          }
        }
      }
    },
...
    "/api/mentors": {
      "get": {
        "tags": ["Mentors"],
        "summary": "List mentors",
        "description": "Returns Mentor[] unless page is provided.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Q" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Track" },
          { "$ref": "#/components/parameters/MentorshipType" },
          { "$ref": "#/components/parameters/Duration" },
          { "$ref": "#/components/parameters/Capacity" },
          { "$ref": "#/components/parameters/Expertise" }
        ],
        "responses": {
          "200": {
            "description": "Mentors",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "type": "array", "items": { "$ref": "#/components/schemas/Mentor" } },
                    { "$ref": "#/components/schemas/PagedMentors" }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Mentors"],
        "summary": "Create mentor",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MentorWrite" } } }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Mentor" } } }
          }
        }
      }
    },
...
    "/api/mentees": {
      "get": {
        "tags": ["Mentees"],
        "summary": "List mentees",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Q" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Track" },
          { "$ref": "#/components/parameters/ApplicationStatus" }
        ],
        "responses": {
          "200": {
            "description": "Mentees",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "type": "array", "items": { "$ref": "#/components/schemas/Mentee" } },
                    { "$ref": "#/components/schemas/PagedMentees" }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Mentees"],
        "summary": "Create mentee",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MenteeWrite" } } }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Mentee" } } }
          }
        }
      }
    },
...
    "/api/groups": {
      "get": {
        "tags": ["Groups"],
        "summary": "List groups",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Q" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Groups",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "type": "array", "items": { "$ref": "#/components/schemas/Group" } },
                    { "$ref": "#/components/schemas/PagedGroups" }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Groups"],
        "summary": "Create group",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GroupWrite" } } }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/activities": {
      "get": {
        "tags": ["Activities"],
        "summary": "List activities",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "Activities",
            "content": {
              "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Activity" } } }
            }
          }
        }
      },
      "post": {
        "tags": ["Activities"],
        "summary": "Create an activity (admin)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Activity" } } }
        },
        "responses": { "201": { "description": "Created" }, "403": { "description": "Forbidden" } }
      }
    },
    "/api/slots": {
      "get": {
        "tags": ["Slots"],
        "summary": "List slots",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "mentorId", "in": "query", "schema": { "type": "string" } },
          { "name": "menteeId", "in": "query", "schema": { "type": "string" } },
          { "name": "availableOnly", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Slots",
            "content": {
              "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Slot" } } }
            }
          }
        }
      },
      "post": {
        "tags": ["Slots"],
        "summary": "Create slot",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SlotWrite" } } }
        },
        "responses": {
          "201": { "description": "Created" },
          "409": {
            "description": "Time conflict",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
    "/api/session-logs": {
      "get": {
        "tags": ["SessionLogs"],
        "summary": "List session logs scoped to the caller",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "mentorId", "in": "query", "schema": { "type": "string" } },
          { "name": "menteeId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Logs",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/SessionLog" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["SessionLogs"],
        "summary": "Create or update a session log",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SessionLogWrite" } } }
        },
        "responses": { "200": { "description": "Saved" } }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Notifications for the signed-in user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Notifications",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NotificationList" } } }
          }
        }
      }
    },
...
    "/api/testimonials": {
      "get": {
        "tags": ["Testimonials"],
        "summary": "List testimonials",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "track", "in": "query", "schema": { "type": "string" } },
          { "name": "q", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Testimonials",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Testimonial" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Testimonials"],
        "summary": "Create a testimonial",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TestimonialWrite" } } }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
  }
}